+++ Complex.hpp
@@ -3,15 +3,15 @@
 class Complex
 {
 private:
-  int real, imag;
+  double real, imag;
 public:
-  Complex(int real, int imag);
+  Complex(double real, double imag);
 
-  int getReal();
-  int getImag();
+  double getReal();
+  double getImag();
 
-  void setReal(int real);
-  void setImag(int imag);
+  void setReal(double real);
+  void setImag(double imag);
 
   friend Complex operator+ (Complex const &comp1, Complex const &comp2);
   friend Complex operator- (Complex const &comp1, Complex const &comp2);
